+++ src/app/maracas/page.tsx
@@ -1,18 +1,11 @@
-"use client";
-import Show3dObj from "@/components/Show3dObj";
-import useSoundHook from "@/customhooks/useSoundHook";
+import Link from "next/link";
 const Page = () => {
-	const { isSoundOn, setIsSoundOn, isDevicemotionPermissionGranted, requestPermission } = useSoundHook();
 	return (
 		<div>
-			<div className='absolute top-0 right-0 z-10'>
-				<button onClick={() => setIsSoundOn(!isSoundOn)}>{isSoundOn ? "Stop" : "Start"} Sound</button>
-				{!isDevicemotionPermissionGranted || true ? <button onClick={requestPermission}>Enable Motion</button> : <></>}
-			</div>
-			<Show3dObj />
+			<Link href={"/maracas/normal"}>ノーマルマラカス</Link>
+			<Link href={"/maracas/special"}>スペシャルマラカス</Link>
 		</div>
 	);
 };
 
 export default Page;
-
